requirements.txt
# Core Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Background Tasks
redis==5.0.1
celery==5.3.6

# ML/NLP
scikit-learn==1.4.0
sentence-transformers==2.3.1
spacy==3.7.2
numpy==1.26.3
pandas==2.2.0

# Resume Parsing
pypdf==4.0.1
python-docx==1.1.0
reportlab==4.0.9

# MLOps
mlflow==2.10.0

# AWS
boto3==1.34.34

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.3
httpx==0.26.0

# Code Quality
black==24.1.1
pylint==3.0.3
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.1
